evaluation: split recorder tests by source file
